requirements.txt
# Install Celery if you are running CKAN 2.6 or less
#celery==3.1.25
progressbar2==3.55.0
#SQLAlchemy>=0.6.6
#requests>=1.1.0
ckantoolkit>=0.0.4
six>=1.12.0

-e git+https://github.com/qld-gov-au/ckanext-report.git@develop#egg=ckanext-report
